Deprecating unused framework/materials. Promoting framework/physics/physics_material to framework/materials. Other minor cleanup in framework/physics.
