--- framework/materials/mat_prop_scalarfuncXYZTV.h
@@ -1,28 +1,0 @@
-// SPDX-FileCopyrightText: 2024 The OpenSn Authors <https://open-sn.github.io/opensn/>
-// SPDX-License-Identifier: MIT
-
-#pragma once
-
-#include "framework/materials/material_property.h"
-#include "framework/math/functions/function_dimA_to_dimB.h"
-
-namespace opensn
-{
-
-/**General material base class for a scalar material property that
- * is possibly a function of position (x,y,z), time t, and maybe a set
- * of variables (v).*/
-class MaterialPropertyScalarFuncXYZTV : public MaterialProperty
-{
-protected:
-  const FunctionDimAToDimB& function_;
-  const std::vector<std::string> dependent_variables_;
-
-public:
-  static InputParameters GetInputParameters();
-  explicit MaterialPropertyScalarFuncXYZTV(const InputParameters& params);
-
-  double Evaluate(const std::vector<double>& vars);
-};
-
-} // namespace opensn
