--- framework/materials/material.cc
@@ -1,29 +1,0 @@
-// SPDX-FileCopyrightText: 2024 The OpenSn Authors <https://open-sn.github.io/opensn/>
-// SPDX-License-Identifier: MIT
-
-#include "framework/materials/material.h"
-
-#include "framework/object_factory.h"
-
-namespace opensn
-{
-
-InputParameters
-Material::GetInputParameters()
-{
-  InputParameters params = Object::GetInputParameters();
-
-  params.AddRequiredParameter<std::string>(
-    "name", "The text name that will be associated with this material.");
-
-  params.AddRequiredParameterArray(
-    "properties", "Expects an array object handles that represents the properties.");
-
-  return params;
-}
-
-Material::Material(const InputParameters& params) : name_(params.GetParamValue<std::string>("name"))
-{
-}
-
-} // namespace opensn
