+++ framework/materials/material_property_isotropic_mg_src.h
@@ -3,18 +3,18 @@
 
 #pragma once
 
-#include "framework/physics/physics_material/material_property_base.h"
+#include "framework/materials/material_property.h"
 
 namespace opensn
 {
 
 /** Basic thermal conductivity material property.*/
-class IsotropicMultiGrpSource : public PhysicsMaterialProperty
+class IsotropicMultiGrpSource : public MaterialProperty
 {
 public:
   std::vector<double> source_value_g_;
 
-  IsotropicMultiGrpSource() : PhysicsMaterialProperty(PropertyType::ISOTROPIC_MG_SOURCE) {}
+  IsotropicMultiGrpSource() : MaterialProperty(PropertyType::ISOTROPIC_MG_SOURCE) {}
 };
 
 } // namespace opensn
