+++ framework/materials/material_property_scalar_value.h
@@ -3,18 +3,18 @@
 
 #pragma once
 
-#include "framework/physics/physics_material/material_property_base.h"
+#include "framework/materials/material_property.h"
 
 namespace opensn
 {
 
 /**Simple scalar material property.*/
-class ScalarValue : public PhysicsMaterialProperty
+class ScalarValue : public MaterialProperty
 {
 public:
   double value_ = 1.0;
 
-  ScalarValue() : PhysicsMaterialProperty(PropertyType::SCALAR_VALUE) {}
+  ScalarValue() : MaterialProperty(PropertyType::SCALAR_VALUE) {}
 
   double GetScalarValue() override { return value_; }
 };
