+++ framework/materials/multi_group_xs/opensn_import.cc
@@ -1,7 +1,7 @@
 // SPDX-FileCopyrightText: 2024 The OpenSn Authors <https://open-sn.github.io/opensn/>
 // SPDX-License-Identifier: MIT
 
-#include "framework/physics/physics_material/multi_group_xs/multi_group_xs.h"
+#include "framework/materials/multi_group_xs/multi_group_xs.h"
 #include "framework/logging/log.h"
 #include <numeric>
