+++ modules/linear_boltzmann_solvers/lbs_solver/lbs_structs.h
@@ -3,8 +3,8 @@
 
 #pragma once
 
-#include "framework/physics/physics_material/multi_group_xs/multi_group_xs.h"
-#include "framework/physics/physics_material/material_property_isotropic_mg_src.h"
+#include "framework/materials/multi_group_xs/multi_group_xs.h"
+#include "framework/materials/material_property_isotropic_mg_src.h"
 #include "framework/math/math.h"
 #include <functional>
 #include <map>
