+++ modules/mg_diffusion/mg_diffusion_solver.h
@@ -4,8 +4,8 @@
 #pragma once
 
 #include "modules/mg_diffusion/mg_diffusion_bndry.h"
-#include "framework/physics/physics_material/multi_group_xs/multi_group_xs.h"
-#include "framework/physics/physics_material/material_property_isotropic_mg_src.h"
+#include "framework/materials/multi_group_xs/multi_group_xs.h"
+#include "framework/materials/material_property_isotropic_mg_src.h"
 #include "framework/physics/solver_base/solver.h"
 #include "framework/math/petsc_utils/petsc_utils.h"
 #include "framework/utils/timer.h"
