+++ test/framework/tutorials/tutorial_06_wdd.cc
@@ -1,5 +1,5 @@
 #include "framework/math/spatial_discretization/finite_volume/finite_volume.h"
-#include "framework/physics/physics_material/multi_group_xs/multi_group_xs.h"
+#include "framework/materials/multi_group_xs/multi_group_xs.h"
 #include "framework/math/quadratures/angular/product_quadrature.h"
 #include "framework/field_functions/field_function_grid_based.h"
 #include "framework/mesh/mesh_continuum/mesh_continuum.h"
